move url normalization to util to use for feeds too
